docs: add references and comments

+++ hands/src/main/java/app/touchlessChef/activity/MainActivity.java
@@ -1,4 +1,4 @@
-package app.touchlessChef.activity.home;
+package app.touchlessChef.activity;
 
 import android.app.Activity;
 import android.app.ActivityOptions;
@@ -28,69 +28,61 @@ import com.google.android.material.navigation.NavigationView;
 import java.util.Objects;
 
 import app.touchlessChef.R;
-import app.touchlessChef.activity.MenuActivity;
 import app.touchlessChef.adapter.DatabaseAdapter;
-import app.touchlessChef.constants.RecipeConstants;
-import app.touchlessChef.fragment.home.ChineseFragment;
-import app.touchlessChef.fragment.home.VietnamFragment;
+import app.touchlessChef.fragment.cuisine.ChineseFragment;
+import app.touchlessChef.fragment.cuisine.VietnamFragment;
 import app.touchlessChef.model.Recipe;
-import app.touchlessChef.fragment.home.BaseFragment;
+import app.touchlessChef.fragment.cuisine.BaseFragment;
 import app.touchlessChef.activity.recipe.CreateRecipeActivity;
 import app.touchlessChef.activity.recipe.ViewRecipeActivity;
-import app.touchlessChef.constants.RecipeEditConstants;
+import app.touchlessChef.constants.RecipeConstants;
+import app.touchlessChef.constants.RecipeConstants.INTENT_KEYS;
+import app.touchlessChef.constants.RecipeConstants.REQUEST_CODES;
 
-public class MainActivity extends MenuActivity implements BaseFragment.FragmentListener {
-    private static final int REQUEST_ADD_RECIPE = 1;
-    private static final int REQUEST_VIEW_RECIPE = 2;
+/**
+ * Reference: https://github.com/aza0092/Cooking-Recipe-Android-App/blob/master/app/src/main/java/ui/MainActivity.java
+ * Modified to use Drawer for Navigation + add default recipe
+ */
+public class MainActivity extends BaseActivity implements BaseFragment.FragmentListener {
     private DatabaseAdapter databaseAdapter;
 
     private ImageView firstView;
     private ImageView secondView;
     private ViewSwitcher mViewSwitcher;
-    private DrawerLayout drawer;
+    private DrawerLayout mDrawer;
+    private NavigationView mNavigationView;
     private String currentCategory;
     private final Activity mActivity = this;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_standard_home_main);
-
+        setContentView(R.layout.activity_main);
 
         Window w = getWindow();
-        w.setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS,
-                WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
+        w.setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
 
         databaseAdapter = DatabaseAdapter.getInstance(this);
+        findViewsByID();
 
-        mToolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(mToolbar);
-        Objects.requireNonNull(getSupportActionBar()).setDisplayShowTitleEnabled(false);
-        CollapsingToolbarLayout mCollapsingToolbarLayout = findViewById(R.id.collapsing_toolbar);
-        firstView = findViewById(R.id.firstView);
-        int defaultImage = R.drawable.vn_botloc;
+        // Set default cuisine
+        int defaultImage = R.drawable.vn_bot_loc;
+        currentCategory = RecipeConstants.VIETNAMESE;
         firstView.setImageResource(defaultImage);
         getSupportFragmentManager().beginTransaction().replace(R.id.fragment_holder,
                 VietnamFragment.newInstance(RecipeConstants.VIETNAMESE)).commit();
-        secondView = findViewById(R.id.secondView);
-        mViewSwitcher = findViewById(R.id.switcher);
 
-        Typeface font = Typeface.createFromAsset(getAssets(), "fonts/Roboto-Medium.ttf");
-        mCollapsingToolbarLayout.setCollapsedTitleTypeface(font);
-        mCollapsingToolbarLayout.setExpandedTitleTypeface(font);
-
-        drawer = findViewById(R.id.drawer_layout);
+        // Drawer
         ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, mToolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.addDrawerListener(toggle);
+                this, mDrawer, mToolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        mDrawer.addDrawerListener(toggle);
         toggle.syncState();
-        NavigationView navigationView = findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(item -> {
+        mNavigationView.setNavigationItemSelectedListener(item -> {
             @DrawableRes int image = -1;
             switch (item.toString()) {
                 case RecipeConstants.VIETNAMESE:
                     currentCategory = RecipeConstants.VIETNAMESE;
-                    image = R.drawable.vn_botloc;
+                    image = R.drawable.vn_bot_loc;
                     getSupportFragmentManager().beginTransaction().replace(R.id.fragment_holder,
                             VietnamFragment.newInstance(RecipeConstants.VIETNAMESE)).commit();
                     break;
@@ -108,15 +100,32 @@ public class MainActivity extends MenuActivity implements BaseFragment.FragmentL
                 firstView.setImageResource(image);
                 mViewSwitcher.showPrevious();
             }
-            drawer.closeDrawer(GravityCompat.START);
+            mDrawer.closeDrawer(GravityCompat.START);
             return true;
         });
     }
 
+    private void findViewsByID() {
+        CollapsingToolbarLayout mCollapsingToolbarLayout = findViewById(R.id.collapsing_toolbar);
+        mToolbar = findViewById(R.id.toolbar);
+        firstView = findViewById(R.id.firstView);
+        secondView = findViewById(R.id.secondView);
+        mViewSwitcher = findViewById(R.id.switcher);
+        mDrawer = findViewById(R.id.drawer_layout);
+        mNavigationView = findViewById(R.id.nav_view);
+
+        // Toolbar attributes
+        setSupportActionBar(mToolbar);
+        Objects.requireNonNull(getSupportActionBar()).setDisplayShowTitleEnabled(false);
+        Typeface font = Typeface.createFromAsset(getAssets(), "fonts/Roboto-Medium.ttf");
+        mCollapsingToolbarLayout.setCollapsedTitleTypeface(font);
+        mCollapsingToolbarLayout.setExpandedTitleTypeface(font);
+    }
+
     @Override
     public void onBackPressed() {
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
+        if (mDrawer.isDrawerOpen(GravityCompat.START)) {
+            mDrawer.closeDrawer(GravityCompat.START);
         } else {
             super.onBackPressed();
         }
@@ -132,8 +141,8 @@ public class MainActivity extends MenuActivity implements BaseFragment.FragmentL
     public boolean onOptionsItemSelected(MenuItem item) {
         if (item.getItemId() == R.id.new_recipe) {
             Intent intent = new Intent(this, CreateRecipeActivity.class);
-            intent.putExtra("category", currentCategory);
-            mActivity.startActivityForResult(intent, REQUEST_ADD_RECIPE);
+            intent.putExtra(INTENT_KEYS.CATEGORY, currentCategory);
+            mActivity.startActivityForResult(intent, REQUEST_CODES.REQUEST_ADD_RECIPE);
         }
         return super.onOptionsItemSelected(item);
     }
@@ -141,53 +150,35 @@ public class MainActivity extends MenuActivity implements BaseFragment.FragmentL
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-        switch (requestCode) {
-            case REQUEST_ADD_RECIPE:
-                switch (resultCode) {
-                    case RecipeEditConstants.RECIPE_ADDED:
-                        Snackbar.make(getWindow().getDecorView(), "Recipe added.",
-                                Snackbar.LENGTH_LONG).show();
-                        break;
-                    case RecipeEditConstants.RECIPE_EDITED:
-                        Snackbar.make(getWindow().getDecorView(), "Recipe modified.",
-                                Snackbar.LENGTH_LONG).show();
-                        break;
-                }
-                break;
-            case REQUEST_VIEW_RECIPE:
-                switch (resultCode) {
-                    case RecipeEditConstants.RECIPE_SHOULD_BE_EDITED:
-                        Recipe recipe = data.getParcelableExtra("recipe");
-                        Intent intent = new Intent(this, CreateRecipeActivity.class);
-                        intent.putExtra("recipe", recipe);
-                        intent.putExtra("category", recipe.getCategory());
-                        intent.putExtra("isEditing", true);
-                        mActivity.startActivityForResult(intent, REQUEST_ADD_RECIPE);
-                        break;
-                    case RecipeEditConstants.RECIPE_SHOULD_BE_DELETED:
-                        long recipeId = data.getLongExtra("recipeId", -1);
-                        if (recipeId != -1) {
-                            onDeleteRecipe(recipeId);
-                        }
-                        break;
-                }
-                break;
+        if (requestCode == REQUEST_CODES.REQUEST_ADD_RECIPE &&
+                resultCode == REQUEST_CODES.RECIPE_ADDED) {
+            Snackbar.make(getWindow().getDecorView(), "Recipe added.",
+                    Snackbar.LENGTH_LONG).show();
+        }
+
+        if (requestCode == REQUEST_CODES.REQUEST_VIEW_RECIPE &&
+                resultCode == REQUEST_CODES.RECIPE_SHOULD_BE_DELETED) {
+            long recipeId = data.getLongExtra(INTENT_KEYS.RECIPE_ID, -1);
+            if (recipeId != -1) {
+                onDeleteRecipe(recipeId);
+            }
+            Snackbar.make(getWindow().getDecorView(), "Recipe deleted.",
+                    Snackbar.LENGTH_LONG).show();
         }
     }
 
     @Override
     public void onShowRecipe(Recipe recipe, Pair<ImageView, String> pairs) {
         Intent intent = new Intent(this, ViewRecipeActivity.class);
-        intent.putExtra("recipe", recipe);
+        intent.putExtra(INTENT_KEYS.RECIPE, recipe);
 
         ActivityOptions transitionActivityOptions =
                 ActivityOptions.makeSceneTransitionAnimation(
                         mActivity, pairs.first, pairs.second);
         mActivity.startActivityForResult(
-                intent, REQUEST_VIEW_RECIPE, transitionActivityOptions.toBundle());
+                intent, REQUEST_CODES.REQUEST_VIEW_RECIPE, transitionActivityOptions.toBundle());
     }
 
-    @Override
     public void onDeleteRecipe(long recipeId) {
         databaseAdapter.deleteRecipe(recipeId);
         Snackbar.make(getWindow().getDecorView(), "Recipe deleted.", Snackbar.LENGTH_LONG).show();
